+++ pkg/core/src/tests/Adapter.t.sol
@@ -18,14 +18,14 @@ contract FakeAdapter is BaseAdapter {
         address _divider,
         address _target,
         address _oracle,
-        uint256 _ifee,
+        uint64 _ifee,
         address _stake,
         uint256 _stakeSize,
-        uint128 _minm,
-        uint128 _maxm,
-        uint8 _mode,
-        uint128 _tilt,
-        uint256 _level
+        uint48 _minm,
+        uint48 _maxm,
+        uint16 _mode,
+        uint64 _tilt,
+        uint16 _level
     ) BaseAdapter(_divider, _target, _oracle, _ifee, _stake, _stakeSize, _minm, _maxm, _mode, _tilt, _level) {}
 
     function _scale() internal virtual override returns (uint256 _value) {
